HTTP視聴時の不具合を修正した
