+++ PeerCastStation/PeerCastStation.Test/PeerCastStation.Test.fsproj
@@ -35,6 +35,8 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Compile Include="TestCommon.fs" />
+    <Compile Include="HttpTests.fs" />
     <Compile Include="Tests.fs" />
     <Compile Include="Program.fs" />
   </ItemGroup>
